Refactor DirectoryViewTest.kt and ProjectEditorControl.kt
